Datamplify-DEV/requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
attrs==25.3.0
boto3==1.40.6
botocore==1.40.6
cassandra-driver==3.29.2
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
clickhouse-connect==0.7.19
cryptography==43.0.1
Django==5.0.7
django-cors-headers==4.7.0
django-oauth-toolkit==3.0.1
django-rest-knox==4.2.0
djangorestframework==3.15.2
dnspython==2.7.0
dotenv==0.9.9
drf-yasg==1.21.10
duckdb==1.1.2
frictionless==5.18.1
geomet==0.2.1.post1
humanize==4.12.3
idna==3.10
inflection==0.5.1
isodate==0.7.2
jmespath==1.0.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jwcrypto==1.5.6
lz4==4.4.4
markdown-it-py==3.0.0
marko==2.2.0
mdurl==0.1.2
numpy==2.2.6
oauthlib==3.3.1
packaging==25.0
pandas==2.2.2
petl==1.7.17
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.9.0
pymongo==4.14.0
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rfc3986==2.0.0
rich==14.1.0
rpds-py==0.27.0
s3transfer==0.13.1
shellingham==1.5.4
simpleeval==1.0.3
six==1.17.0
tabulate==0.9.0
text-unidecode==1.3
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
validators==0.35.0
zstandard==0.23.0
SQLAlchemy<2.0
